Specify the useExternalCss config into the README.md
The new configuration variable has been added to the README.md

By default, it is set to false.
If set to true, the local tarteaucitron.css file will not be loaded.

The name of the config might not be self explanatory. I would have recommended changing it to "loadLocalCss". (with true by default)